requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sse-starlette>=2.0.0
httpx>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyMuPDF>=1.23.0
Pillow>=10.2.0
pillow-heif>=0.14.0
openpyxl>=3.1.2
python-magic>=0.4.27
google-genai>=1.0.0
openai>=1.12.0
anthropic>=0.18.0
